refactor: extract body parser, tree renderer, post-processor, and split types
- Extract body-parser.ts to core/curl (consolidate JSON parsing from 3 executors)
- Extract tree-renderer.ts to core/format (reusable tree output)
- Extract post-processor.ts to core/execution (validation + snapshot logic)
- Split types/config.ts into domain modules (json, request, execution, snapshot, diff, profile, watch, global)
- Consolidate ANSI colors usage in diff-formatter and snapshot-formatter
- Add 14 new tests for post-processor
