+++ packages/cli/src/core/execution/index.ts
@@ -3,5 +3,7 @@
  * Provides retry logic and execution helpers.
  */
 
+export type { PostProcessOptions } from './post-processor';
+export { postProcessResult } from './post-processor';
 export { calculateBackoffDelay, createRetryOptions, withRetry } from './retry';
 export type { RetryConfig, RetryLogger, RetryOptions, RetryResult } from './types';
