+++ packages/cli/src/types/json.ts
@@ -1,0 +1,11 @@
+/**
+ * JSON primitive types for request/response bodies.
+ */
+
+export type JsonValue = string | number | boolean | null | JsonObject | JsonArray;
+
+export interface JsonObject {
+  [key: string]: JsonValue;
+}
+
+export interface JsonArray extends Array<JsonValue> {}
